requirements.txt
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
fpdf==1.7.2
idna==2.9
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.1.2
pylint==2.5.2
PyMySQL==0.9.3
python-barcode==0.11.0
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
urllib3==1.25.9
wrapt==1.12.1
